Create `UserObjectConverter` class
